#10076: Merge cos_bw to TTNN
