+++ docs/source/ttnn/ttnn/dependencies/tt_lib.rst
@@ -848,8 +848,6 @@ Backward Operations
 
 .. autofunction:: tt_lib.tensor.cosh_bw
 
-.. autofunction:: tt_lib.tensor.cos_bw
-
 .. autofunction:: tt_lib.tensor.acosh_bw
 
 .. autofunction:: tt_lib.tensor.acos_bw
