+++ docs/source/ttnn/ttnn/ttnn/cos_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.cos_bw:
+
+ttnn.cos_bw
+############
+
+.. autofunction:: ttnn.cos_bw
